requirements.txt
streamlit>=1.31.1
openai>=0.27.0
python-dotenv>=0.19.0
requests>=2.31.0
altair>=4.2.2 